change MealPlanDay and MealPlanDayFactory to follow factory pattern
