+++ src/Entities/MealPlanDayFactory.java
@@ -1,7 +1,36 @@
 package Entities;
 
-public class MealPlanDayFactory {
-    public MealPlanDay createMealPlanDay(Recipe breakfastRecipe, Recipe lunchRecipe, Recipe dinnerRecipe) {
-        return new MealPlanDay(breakfastRecipe, lunchRecipe, dinnerRecipe);
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public class MealPlanDayFactory implements MealPlanDay {
+    private final Recipe breakfastRecipe;
+    private final Recipe lunchRecipe;
+    private final Recipe dinnerRecipe;
+
+    @JsonCreator
+    public MealPlanDayFactory(@JsonProperty("breakfastRecipe") Recipe breakfastRecipe, @JsonProperty("lunchRecipe") Recipe lunchRecipe, @JsonProperty("dinnerRecipe") Recipe dinnerRecipe) {
+        this.breakfastRecipe = breakfastRecipe;
+        this.lunchRecipe = lunchRecipe;
+        this.dinnerRecipe = dinnerRecipe;
     }
+
+    public MealPlanDayFactory() {
+        breakfastRecipe = null;
+        lunchRecipe = null;
+        dinnerRecipe = null;
+    }
+
+    public Recipe getBreakfastRecipe(){
+        return breakfastRecipe;
+    }
+
+    public Recipe getLunchRecipe(){
+        return lunchRecipe;
+    }
+
+    public Recipe getDinnerRecipe(){
+        return dinnerRecipe;
+    }
+
 }
